Refactor TerminalClones into it's own file
